MaverickService: Improvements
All portions of the service are now done system-level so app only handles settings now.
